Replaced github client with era contracts repo
