+++ cli/package.json
@@ -45,6 +45,7 @@
     "chalk": "^5.3.0",
     "cli-table3": "^0.6.4",
     "dotenv": "^16.4.5",
+    "nochoices": "^1.1.4",
     "ora": "^8.0.1",
     "simple-git": "^3.24.0",
     "tempy": "^3.1.0",
