+++ cli/src/lib/era-contracts-repo.ts
@@ -7,18 +7,22 @@ import { promisify } from "node:util";
 import { compiledArtifactParser } from "../schema/compiled";
 import { utils } from "zksync-ethers";
 import { systemContractHashesSchema } from "../schema/github-schemas";
+import type { Sources } from "../schema";
+import { Option } from "nochoices";
 
 const execPromise = promisify(exec)
 
 export class EraContractsRepo {
   repoPath: string;
   git: SimpleGit;
+  private _currentRef: Option<Promise<string>>;
 
   constructor(repoPath: string) {
     this.repoPath = repoPath;
     this.git = simpleGit({
       baseDir: repoPath,
     });
+    this._currentRef = Option.None()
   }
 
   static async default(): Promise<EraContractsRepo> {
@@ -36,11 +40,12 @@ export class EraContractsRepo {
     }
   }
 
-  async readFile(subPath: string, ref: string): Promise<string> {
+  async readFile(subPath: string, ref = 'HEAD'): Promise<string> {
     return this.git.show(`${ref}:${subPath}`);
   }
 
-  async setRevision(ref: string):Promise<void> {
+  async setRevision(ref: string): Promise<void> {
+    this._currentRef.insert(Promise.resolve(ref))
     await this.git.checkout(ref, ['--force'])
   }
 
@@ -60,8 +65,19 @@ export class EraContractsRepo {
       throw new Error(`Unknown contract: ${systemContractName}`)
     }
 
-    const contractArtifactoryFile = path.join(this.repoPath, "system-contracts", hashData.bytecodePath)
+    return this.extractBytecodeFromFile(hashData.bytecodePath)
+  }
 
+  async byteCodeHashFor(systemContractName: string): Promise<string> {
+    const byteCode = await this.byteCodeFor(systemContractName)
+    const rawHash = utils.hashBytecode(byteCode);
+    const hex = Buffer.from(rawHash).toString('hex');
+    return `0x${hex}`
+  }
+
+
+  private async extractBytecodeFromFile(filePath: string): Promise<Buffer> {
+    const contractArtifactoryFile = path.join(this.repoPath, "system-contracts", filePath)
     if (contractArtifactoryFile.endsWith('.json')) {
       const content = await fs.readFile(contractArtifactoryFile)
       const json = JSON.parse(content.toString())
@@ -74,10 +90,44 @@ export class EraContractsRepo {
     }
   }
 
-  async byteCodeHashFor(systemContractName: string): Promise<string> {
-    const byteCode = await this.byteCodeFor(systemContractName)
-    const rawHash = utils.hashBytecode(byteCode);
-    const hex = Buffer.from(rawHash).toString('hex');
-    return `0x${hex}`
+  async downloadSystemContract(contractName: string): Promise<Sources> {
+    return this.downloadContractInt(`system-contracts/contracts/${contractName}.sol`, {});
+  }
+
+  private async downloadContractInt(rootPath: string, partial: Sources): Promise<Sources> {
+    if (partial[rootPath]) {
+      return partial;
+    }
+
+    const dir = path.parse(rootPath).dir;
+    const content = await this.readFile(rootPath);
+    partial[rootPath] = { content };
+
+    const deps = this.extractDeps(content);
+
+    await Promise.all(
+      deps.map(async (dep: string): Promise<void> => {
+        const depPath = path.normalize(path.join(dir, dep));
+        await this.downloadContractInt(depPath, partial);
+      })
+    );
+
+    return partial;
+  }
+
+  private extractDeps(sourceCode: string) {
+    const reg1 = /import.*["'](.+)["'];?/g;
+    const matches = sourceCode.matchAll(reg1);
+
+    const res: string[] = [];
+    for (const match of matches) {
+      res.push(match[1]);
+    }
+
+    return res;
+  }
+
+  async currentRef(): Promise<string> {
+    return this._currentRef.getOrInsert(this.git.revparse('--short HEAD'))
   }
 }
